refactor: 拆分模块结构，新增资源文件、测试和工具
- 从 lib/index.js 提取常量到 lib/constants.js、组件到 lib/widget.js
- 新增角色图片素材 (DSniang) 和音效文件
- 新增 test/ 和 tools/ 目录
- 更新 README、package.json、cordis.patch.yml 等配置文件
- 新增 CHANGELOG.md
