+++ CHANGELOG.md
@@ -1,0 +1,13 @@
+# 更新日志 / Changelog
+
+本项目遵循 [语义化版本](https://semver.org/lang/zh-CN/)。
+
+## 1.0.0
+
+首次发布。
+
+- 基于 [MeteorNOX/DeepSeek-Balance-Whale-Widget](https://github.com/MeteorNOX/DeepSeek-Balance-Whale-Widget) 
+- 形象菜单四选一：默认 / 奶鲸 / 糖鲸 / 睡觉
+- 奶鲸形象已压缩至 310KB（原图 4MB → 310KB，压缩 92.4%）
+- 路由前缀改为 `/dsh-whale-naijing/`，插件 id 改为 `dsh-whale-naijing`
+- 可与原版共存
