+++ cordis.patch.yml
@@ -1,15 +1,13 @@
 # ============================================================================
-# dsh-whale-widget —— DSH bundle patch
+# dsh-whale-naijing —— DSH bundle patch
 # ============================================================================
 # 这是插件的"挂载声明"。DSH 启动时按 bundle 层栈叠加各插件 patch，
 # 本文件会把鲸鱼余额挂件插入 Web profile 的配置树。
 #
-# 安装方式（不再需要手工复制/手工改 profile patch）：
-#   dsh plugin --profile web add dsh-whale-widget
-# 或本地开发安装：
+# 安装方式：
 #   dsh plugin --profile web add link:<本目录绝对路径>
 # ============================================================================
 
 - insert:
-    - id: dsh-whale-widget
-      name: dsh-whale-widget
+    - id: dsh-whale-naijing
+      name: dsh-whale-naijing
