+++ tools/README.md
@@ -1,0 +1,17 @@
+## 使用说明
+
+直接双击打开 `zzz-position.html`，在浏览器里调整：
+
+### 操作方式
+1. **拖动滑块** — 水平（left%）和垂直（top%）实时预览
+2. **点击容器** — 直接在挂件区域内点击，ZZZ 跟随鼠标
+3. **十字参考线** — 黄色虚线实时显示坐标轴，方便对齐角色头顶
+4. **复制 CSS** — 调到满意后点「📋 复制最终 CSS」，粘贴给我即可
+
+### 参考坐标
+- 角色图（睡觉）：占据容器 **右下角**（right:0, bottom:0, 59.45%）
+- 角色头顶大约在：**left ~70% / top ~41%**
+- 气泡文字：**left:44.25% / top:38%**
+
+### 输出格式
+点复制后会得到一行完整 CSS，直接发给我，我会更新到代码里。
